requirements.txt
fastapi==0.115.0
uvicorn==0.22.0
python-dotenv==1.0.1
httpx==0.24.1
pymysql==1.1.0
aiomysql==0.2.0
requests==2.32.3
sshtunnel==0.4.0  
paramiko==3.4.0 
jinja2==3.1.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.6
bcrypt==4.0.1
email-validator==2.3.0
motor==3.7.1
fastapi-mail==1.6.1
pydantic==2.12.5
pydantic-settings==2.12.0
certifi==2024.8.30
